ref-code-backend/requirements.txt
google-cloud-aiplatform==1.92.0
uvicorn==0.34.0
PyMuPDF==1.24.1.0
google-cloud-storage==2.19.0
google-genai==1.14.0
gitpython==3.1.40
google-adk==0.5.0
deprecated
fastapi==0.115.0
python-multipart==0.0.12
pydantic==2.9.0
email-validator>=2.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
google-auth>=2.26.1
google-auth-oauthlib==1.1.0
google-auth-httplib2>=0.1.1
psycopg2-binary>=2.9.9
PyJWT==2.8.0
PyYAML>=6.0.1
aiohttp>=3.9.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx>=0.28.1,<1.0.0
pytest-cov==4.1.0
